iteration and cutting overhaul
Streamlines cluster_cutter to avoid code repetition, and changed
the generation of sequences fed to it to hopefully make it easier
to follow how sequences are sliced, depending on the strand.

Added a column with the gene id, and changed the logic so that
the "start" columns are always lower than the "end" columns (should
be easier to follow for users).

Added an option to change the meaning of the `--downstream` arg;
if the user provides `--downstream-start-codon`, then the
downstream positions are relative to the gene's start codon.