Add the option to stop execution if anything is missing
Samples, genes or chromosomes might be missing from the inputs;
so far panfeed would throw a warning and continue.
This change introduces an option to stop execution if anything
is missing.

This change also introduces a check on the presence of a chromosome,
which may be missing from the assembly but indicated in the gff
file.